Business functions

+++ main/src/main/resources/db.changelog/04.create-types_of_container-table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE types_of_container (
     number_of_calories INTEGER NOT NULL,
     name VARCHAR(30) NOT NULL,
-    price INTEGER NOT NULL,
+    price DECIMAL NOT NULL,
     CONSTRAINT typesofcontainer_pkey PRIMARY KEY (number_of_calories),
     CONSTRAINT typesofcontainer_numberofcalories_key UNIQUE (number_of_calories),
     CONSTRAINT typesofcontainer_price_key UNIQUE (price)
